Threads now run constantly without being stopped / restarted
Eliminated some unnecessary code

+++ scrape_drive.py
@@ -14,8 +14,6 @@ import time
 
 # ============== Default Property Values ==============
 
-begin_scrape = False
-
 # ---------- Customizable paths
 # URL of Assemblyline instance
 al_instance = 'https://134.190.171.253/'
@@ -48,18 +46,12 @@ list_to_receive = []
 pass_files = []
 # List holds all potentially malicious files as determined by AL output
 mal_files = []
-# True if our submitting thread is currently active
-submitting = False
-# True if our receiving thread is currently active
-receiving = False
-# True is results have already been output
-results = False
+
+scrape_stage = 0
 
 # ---------- Kiosk communication
 # Creates socket between web app and this module
 socketIO = SocketIO
-# True when files are currently being uploaded
-loading = False
 
 
 # ============== Helper Functions ==============
@@ -70,10 +62,51 @@ def kiosk(msg):
     :param msg: message to be sent
     :return:
     """
+
     global socketIO
 
     print msg
+
     socketIO.emit('to_kiosk', msg)
+
+
+def initialize():
+    """
+    Constantly running loop which watches our ingest directory for changes. Any files that are copied into this
+    directory are submitted to AL server
+    :return:
+    """
+
+    global terminal
+
+    print "init_workers"
+
+    refresh_socket()
+    refresh_session()
+
+    # Connects to our Assemblyline deployment
+    terminal = Client(al_instance, apikey=('admin', 'CbHIT^4L*SqLUOoNwLE5g67TaRL9IZEnmE*omXHIC8AI(G3q'), verify=False)
+
+    # 1. Initializes pyudev observer thread that is going to monitor for device events (devices added / removed)
+    monitor = pyudev.Monitor.from_netlink(context)
+    observer = pyudev.MonitorObserver(monitor, block_event)
+    observer.start()
+
+    # 2. Initializes submit thread. Takes files added to list_to_submit array and submits them to AL server
+    st = Thread(target=submit_thread, args=('ingest_queue',), name="submit_thread")
+    st.daemon = True
+    st.start()
+
+    # 3. Initializes receive thread. This thread listens for callbacks from the AL server
+    rt = Thread(target=receive_thread, args=('ingest_queue',), name="receive_thread")
+    rt.daemon = True
+    rt.start()
+
+    # 4. Begins infinite loop that watches for files being added to the ingest_dir folder. Files are copied over when
+    #    detected by the observer; this loop detects the new file, and adds its path to the list_to_submit array to be
+    #    picked up by the submit thread
+    dt = Thread(target=detect_thread, name="detect_thread")
+    dt.start()
 
 
 def refresh_socket():
@@ -81,7 +114,10 @@ def refresh_socket():
     Refreshes our socketio connection
     :return:
     """
+
     global socketIO
+
+    print "refresh_socket"
 
     try:
         socketIO = SocketIO('http://10.0.2.2:5000', verify=False)
@@ -89,90 +125,22 @@ def refresh_socket():
         time.sleep(2)
         refresh_socket()
 
-
-def check_done():
-    """
-    Checks whether or not all files have been imported from our device
-    :return:
-    """
-    global loading
-    global socketIO
-    global results
-    global partition_toread
-    global list_to_submit
-    global terminal_id
-
-    time.sleep(2)
-
-    # Checks if all partitions have been mounted, all files from partitions have been ingested, and all our ingested
-    # files have returned messages from the server. If all these are true then we are finished ingesting files and
-    # our submit and receive threads are shut down. Once lists have been emitted they are reset.
-    if partition_toread == 0 and len(list_to_submit) == 0 and not results:
-        results = True
-        kiosk('\n--- All files have been successfully ingested')
-        if loading:
-            socketIO.emit('device_event', 'done_loading')
-            loading = False
-        time.sleep(0.1)
-        kiosk("\r\n")
-        time.sleep(0.1)
-        socketIO.emit('pass_files', pass_files)
-        time.sleep(0.1)
-        socketIO.emit('mal_files', mal_files, terminal_id)
-        time.sleep(0.1)
+    print "refresh_socket done"
 
 
-def init_ingester():
-    """
-    Constantly running loop which watches our ingest directory for changes. Any files that are copied into this
-    directory are submitted to AL server
-    :return:
-    """
-
-    monitor = pyudev.Monitor.from_netlink(context)
-    observer = pyudev.MonitorObserver(monitor, block_event)
-    observer.start()
-
-    # Sets up inotify to watch imported_files directory
-    i = adapters.InotifyTree(ingest_dir)
-
-    # Loop watches for new additions to imported_files directory
-    for event in i.event_gen():
-        if event is not None:
-            # Stores the event type, pathname, and filename for this event
-            (_, type_names, path, filename) = event
-            for e_type in type_names:
-                # If our event is that we've finished writing a file to imported_files, passes that file's path into
-                # our list_to_submit
-                if e_type == 'IN_CLOSE_WRITE' and filename != '':
-                    dir_to_ingest = path + '/' + filename
-                    socketIO.emit("ingest_status", "submit_file")
-                    list_to_submit.append(dir_to_ingest)
-                    if not submitting:
-                        st = Thread(target=submit_thread, args=('ingest_queue',), name="submit_thread")
-                        st.daemon = True
-                        st.start()
-
-
-def init_session():
-    global terminal
+def refresh_session():
     global list_to_submit
     global list_to_receive
     global mal_files
     global pass_files
-    global results
 
-    print "init_session"
-
-    # Connects to our Assemblyline deployment
-    terminal = Client(al_instance, apikey=('admin', 'CbHIT^4L*SqLUOoNwLE5g67TaRL9IZEnmE*omXHIC8AI(G3q'), verify=False)
+    print "refresh_session"
 
     # Resets all default values
     list_to_submit = []
     list_to_receive = []
     mal_files = []
     pass_files = []
-    results = False
 
 
 def app_response(f_name, l_name, default_settings):
@@ -182,29 +150,59 @@ def app_response(f_name, l_name, default_settings):
     are allowed to run (ie. the service starts as normal)
     :param f_name: Client first name
     :param l_name: Client last name
+    :param default_settings: User settings passed down by application
     :return:
     """
 
-    global begin_scrape
+    global scrape_stage
+
     if f_name != '' and l_name != '':
-        begin_scrape = True
-        print f_name
-        print l_name
-        print default_settings
+        scrape_stage = 1
+        print "  User First Name: " + f_name
+        print "  User Last Name: " + l_name
 
 
 def session_wait():
 
+    global scrape_stage
     print "session_wait"
 
-    refresh_socket()
+    scrape_stage = 0
 
-    while not begin_scrape:
+    while scrape_stage == 0:
         socketIO.emit("connect_request", app_response)
         socketIO.wait_for_callbacks(seconds=1)
         time.sleep(1)
 
-    init_session()
+
+def check_done():
+    """
+    Checks whether or not all files have been imported from our device
+    :return:
+    """
+    global socketIO
+    global scrape_stage
+
+    # Checks if all partitions have been mounted, all files from partitions have been ingested, and all our ingested
+    # files have returned messages from the server. If all these are true then we are finished ingesting files and
+    # our submit and receive threads are shut down. Once lists have been emitted they are reset.
+    # print
+    # print "partition_toread: " + str(partition_toread)
+    # print "list_to_submit: " + str(len(list_to_submit))
+    # print "list_to_receive: " + str(len(list_to_receive))
+    # print
+    if partition_toread == 0 and len(list_to_submit) == 0 and len(list_to_receive) == 0 and scrape_stage == 2:
+        scrape_stage = 3
+        print "Threads: " + str(threading.activeCount())
+        kiosk('\n--- All files have been successfully ingested')
+        socketIO.emit('device_event', 'done_loading')
+        time.sleep(0.1)
+        kiosk("\r\n")
+        time.sleep(0.1)
+        socketIO.emit('pass_files', pass_files)
+        time.sleep(0.1)
+        socketIO.emit('mal_files', mal_files, terminal_id)
+        time.sleep(0.1)
 
 
 # ============== Backend Functions ==============
@@ -234,8 +232,6 @@ def block_event(action, device):
             # Announces a new device has been detected
             if device.get('DEVTYPE') == 'disk':
                 socketIO.emit('device_event', 'connected')
-                time.sleep(0.1)
-                kiosk('\n--- New block device detected: ' + device_id)
 
                 # Makes new folder to hold partitions from this disk
                 path_new = os.path.normpath(ingest_dir + device_id)
@@ -254,8 +250,6 @@ def block_event(action, device):
     # Called when the active device is removed. Clears the imported cart files
     elif action == 'remove':
         if device.get('DEVTYPE') == 'partition':
-            kiosk('Partition removed')
-            kiosk('\nRemoving temp files')
             clear_files(device.device_node)
 
 
@@ -274,7 +268,7 @@ def copy_files(device_id):
     global mount_dir
     global ingest_dir
 
-    while True and len(active_devices) != 0:
+    while len(active_devices) != 0:
 
         # Waits until mount_lock is available (ie. other partitions have finished mounting)
         with mount_lock:
@@ -285,13 +279,6 @@ def copy_files(device_id):
                 # Mounts device
                 os.system('sudo /home/user/al_ui/bash_scripts/mount_block.sh ' + device_id +
                           ' ' + mount_dir)
-
-                # Takes image of device firmware. This is commented out as there are currently no services in
-                # assemblyline that read img files:
-                # os.system('sudo /home/user/al_ui/bash_scripts/image_device_firmware.sh ' + device_id +
-                # ' /home/user/al_ui/temp_device/usb_firm.img')
-
-                kiosk('\n--- Partition successfully loaded: ' + device_id)
 
                 # Makes new directory for this partition
                 os.system('mkdir -p ' + ingest_dir + device_id)
@@ -324,7 +311,6 @@ def neuter_files(raw_dir, device_id):
     """
 
     # Walks through files in given directory one by one, converting to CART and sending to imported_files
-    kiosk('Uploading files...')
     for root, dirs, files in os.walk(raw_dir):
         for raw_file in files:
             output_file = ingest_dir + device_id + '/' + raw_file + '.cart'
@@ -341,7 +327,6 @@ def clear_files(device_id):
     global list_to_submit
     global mal_files
     global partition_toread
-    global loading
     global socketIO
 
     if len(active_devices) != 0:
@@ -349,20 +334,34 @@ def clear_files(device_id):
 
     # Removes the folder in imported_files containing the files corresponding to the removed device
     os.system('rm -rf ' + ingest_dir + device_id)
-    kiosk('Temp files successfully removed: ' + device_id)
 
     # If all devices have been removed, resets list and clears the imported files directory
     if len(active_devices) == 0:
         os.system('rm -rf ' + ingest_dir + '/*')
-        loading = False
-        kiosk('Temporary malicious files have been cleared')
-        kiosk('\n--- All devices successfully removed')
-        time.sleep(0.1)
+        refresh_session()
+        session_wait()
         socketIO.emit('device_event', 'disconnected')
-        kiosk("\r\n")
+        print "Device Removed"
 
 
 # ============== AL Server Interaction Threads ==============
+
+def detect_thread():
+    i = adapters.InotifyTree(ingest_dir)
+    # Loop watches for new additions to imported_files directory
+    for event in i.event_gen():
+        if event is not None:
+            # Stores the event type, pathname, and filename for this event
+            (_, type_names, path, filename) = event
+            for e_type in type_names:
+                # If our event is that we've finished writing a file to imported_files, passes that file's path into
+                # our list_to_submit
+                if e_type == 'IN_CLOSE_WRITE' and filename != '':
+                    dir_to_ingest = path + '/' + filename
+                    socketIO.emit("ingest_status", "submit_file")
+                    list_to_submit.append(dir_to_ingest)
+                    print "        Push: " + dir_to_ingest
+
 
 def submit_thread(queue):
     """
@@ -373,52 +372,47 @@ def submit_thread(queue):
     """
 
     global list_to_submit
-    global loading
     global terminal
     global socketIO
-
-    global submitting
-    global receiving
     global list_to_receive
     global terminal_id
-
-    submitting = True
+    global scrape_stage
 
     # Continuously monitors the list_to_submit. If a new entry is detected, uploads to server and deletes once done
-    while len(list_to_submit):
+    while True:
 
-        if not loading:
-            socketIO.emit('device_event', 'loading')
-            loading = True
+        if len(list_to_submit) and scrape_stage > 0:
 
-        # Pops a file path from the list of files to be submitted
-        ingest_path = list_to_submit.pop()
+            if scrape_stage != 2:
+                socketIO.emit('device_event', 'loading')
+                scrape_stage = 2
+                time.sleep(2)
 
-        # Checks to make sure the file at this path still exists
-        if os.path.exists(ingest_path):
+            # Pops a file path from the list of files to be submitted
+            ingest_path = list_to_submit.pop()
+            print "        Pop: " + ingest_path
 
-            # Checks if the file is empty; ingest is unable to examine empty files. Returns a warning if one is
-            # submitted
-            if os.stat(ingest_path).st_size != 0:
+            # Checks to make sure the file at this path still exists
+            if os.path.exists(ingest_path):
 
-                # Ingests file. Removes file from terminal once ingested
-                kiosk('Ingesting: ' + os.path.basename(ingest_path))
-                list_to_receive.append(os.path.basename(ingest_path))
+                # Checks if the file is empty; ingest is unable to examine empty files. Returns a warning if one is
+                # submitted
+                if os.stat(ingest_path).st_size != 0:
 
-                if not receiving:
-                    rt = Thread(target=receive_thread, args=('ingest_queue',), name="receive_thread")
-                    rt.daemon = True
-                    rt.start()
+                    # Ingests file. Removes file from terminal once ingested
+                    kiosk('Ingesting: ' + os.path.basename(ingest_path))
+                    list_to_receive.append(os.path.basename(ingest_path))
 
-                terminal.ingest(ingest_path,
-                                metadata={'path': ingest_path, 'filename': os.path.basename(ingest_path)},
-                                nq=queue, ingest_type=terminal_id)
-                os.system('rm -f \'' + ingest_path + '\'')
+                    terminal.ingest(ingest_path,
+                                    metadata={'path': ingest_path, 'filename': os.path.basename(ingest_path)},
+                                    nq=queue, ingest_type=terminal_id)
+                    os.system('rm -f \'' + ingest_path + '\'')
 
-            else:
-                kiosk('Unable to ingest, empty file: ' + os.path.basename(ingest_path))
+                else:
+                    kiosk('Unable to ingest, empty file: ' + os.path.basename(ingest_path))
 
-    submitting = False
+        else:
+            time.sleep(1)
 
 
 def receive_thread(queue):
@@ -434,49 +428,45 @@ def receive_thread(queue):
     global mal_files
     global partition_toread
     global list_to_submit
-    global loading
     global active_devices
     global terminal
     global socketIO
 
-    global receiving
+    while True:
 
-    receiving = True
+        if len(list_to_receive):
+            # Takes all messages from the Assemblyline server and stores in list
+            msgs = terminal.ingest.get_message_list(queue)
 
-    while len(list_to_receive):
+            # For each new message that comes from our Assemblyline server, outputs some info about that file. Any files
+            # with a score over 500 have their sid added to the mal_files list. We subtract 1 from num_waiting each time
+            # a result is output
+            for msg in msgs:
+                new_file = os.path.basename(msg['metadata']['path'])
+                if new_file in list_to_receive:
+                    list_to_receive.remove(new_file)
+                    socketIO.emit("ingest_status", "receive_file")
 
-        # Takes all messages from the Assemblyline server and stores in list
-        msgs = terminal.ingest.get_message_list(queue)
+                    score = msg['metadata']['al_score']
+                    sid = msg['alert']['sid']
+                    kiosk('   Server Received: ' + new_file)
 
-        # For each new message that comes from our Assemblyline server, outputs some info about that file. Any files
-        # with a score over 500 have their sid added to the mal_files list. We subtract 1 from num_waiting each time
-        # a result is output
-        for msg in msgs:
-            new_file = os.path.basename(msg['metadata']['path'])
-            if new_file in list_to_receive:
-                list_to_receive.remove(new_file)
-                socketIO.emit("ingest_status", "receive_file")
+                    if score >= 500:
+                        kiosk('        [ ! ] WARNING - Potentially malicious file: ' + new_file)
+                        full_msg = terminal.submission.full(sid)
+                        full_path = full_msg['submission']['metadata']['path']
+                        full_msg['submission']['metadata']['path'] = full_path[full_path.find('temp_device') + 11:]
+                        mal_files.append(full_msg)
 
-                score = msg['metadata']['al_score']
-                sid = msg['alert']['sid']
-                # kiosk('   Server Received: ' + new_file + "    " + 'sid: %s    score: %d' % (sid, score),)
-                kiosk('   Server Received: ' + new_file)
+                    elif score > 0:
+                        full_path = msg['metadata']['path']
+                        msg['metadata']['path'] = full_path[full_path.find('temp_device') + 11:]
+                        pass_files.append(msg)
+                        print
 
-                if score >= 500:
-                    kiosk('        [ ! ] WARNING - Potentially malicious file: ' + new_file)
-                    full_msg = terminal.submission.full(sid)
-                    full_path = full_msg['submission']['metadata']['path']
-                    full_msg['submission']['metadata']['path'] = full_path[full_path.find('temp_device') + 11:]
-                    mal_files.append(full_msg)
-
-                elif score > 0:
-                    full_path = msg['metadata']['path']
-                    msg['metadata']['path'] = full_path[full_path.find('temp_device') + 11:]
-                    pass_files.append(msg)
-                    print
-
-    receiving = False
-    check_done()
+        else:
+            check_done()
+            time.sleep(1)
 
 
 # ============== Initialization ==============
@@ -486,5 +476,5 @@ if __name__ == '__main__':
     os.system('mkdir -p ' + mount_dir)
     os.system('mkdir -p ' + ingest_dir)
 
+    initialize()
     session_wait()
-    init_ingester()
